requirements.txt
requests
pyyaml
nbtlib